chore: improve chat modal trigger styles
